requirements.txt
torch=2.7.1
torchvision=0.22.1
pandas=2.3.1
numpy=2.3.1
tqdm=4.67.1
pyyaml=6.0.2
scikit-learn=1.7.1
matplotlib=3.10.3